requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.3.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==6.1.0
blinker==1.4
bloom==0.12.0
breezy==3.2.1
Brlapi==0.8.3
Brotli==1.0.9
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2024.7.4
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.18.4
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-mixin==0.2.3
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
contourpy==1.2.1
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.8
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.2
dlib==19.24.6
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
dulwich==0.20.31
duplicity==0.8.21
editdistance==0.8.1
empy==3.3.4
exceptiongroup==1.2.1
executing==2.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.112.2
fastbencode==0.0.5
fasteners==0.14.1
fastimport==0.9.14
fastjsonschema==2.19.1
filelock==3.15.4
flake8==4.0.1
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
fs==2.4.12
fsspec==2024.6.1
future==0.18.2
gast==0.5.2
glfw==2.8.0
glm==0.4.4
gpg==1.16.0
greenlet==3.0.3
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.3
image==1.5.33
importlib-metadata==4.6.4
iniconfig==1.1.1
ipykernel==6.29.4
ipython==8.12.3
ipython-genutils==0.2.0
ipython-sql==0.5.0
ipywidgets==8.1.2
isoduration==20.11.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
lark==1.1.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mccabe==0.6.1
mediapipe==0.10.18
mercurial==6.1.1
mistune==3.0.2
ml_dtypes==0.5.0
monotonic==1.6
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==0.0.0
mutagen==1.45.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
nose2==0.9.2
notebook==7.2.0
notebook_shim==0.2.4
notify2==0.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
patiencediff==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.5.0
platformdirs==4.2.1
pluggy==0.13.0
ply==3.11
prettytable==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.5
psutil==5.9.8
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
py-cpuinfo==9.0.0
py-spy==0.3.14
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
pygame==2.6.1
PyGithub==1.55
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gitlab==2.10.1
python-json-logger==2.0.7
python-multipart==0.0.9
pythran==0.10.0
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
referencing==0.35.1
reportlab==3.6.8
requests==2.31.0
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman==3.3
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rospkg-modules==1.6.0
rpds-py==0.18.0
scipy==1.14.1
seaborn==0.13.2
SecretStorage==3.3.1
selenium==4.23.1
Send2Trash==1.8.3
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.38.2
sympy==1.9
systemd-python==234
terminado==0.18.1
texttable==1.6.4
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.3
transforms3d==0.4.2
trio==0.26.2
trio-websocket==0.11.1
triton==3.0.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
ultralytics==8.2.83
ultralytics-thop==2.0.5
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.30.6
variety==0.8.5
vcstool==0.3.0
vcstools==0.1.42
virtualenv==20.27.0
wadllib==1.3.6
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==9.1
widgetsnbextension==4.0.10
wrapt==1.13.3
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarg==0.1.9
yt-dlp==2022.4.8
zipp==1.0.0
